Fix: provide an array where the code expects one
The function output_form in categories.class.php expects an array in $form_input. When the function get called to create the form for a new category, the function call provided an emptry string. This led to a warning since PHP 5.4 because of type mismatch. Providing an array prevents this error. This is the only pace, I found the error.